refactor(engineer_worktree): extract sweep_orphaned_worktrees + helpers (#1266)
Continue mod.rs splitting after #1366: extract sweep_orphaned_worktrees,
git_capture, validate_goal_id, is_valid_sha40 into sibling sweep.rs.

mod.rs: 650 -> 436 (still slightly over 400, last small split next)
sweep.rs: 228 LOC

All 17 engineer_worktree tests pass.
